requirements.txt
# ============================================
# CORE & CLI
# ============================================
click==8.3.1
rich==14.3.3
colorama>=0.4.6
colorama==0.4.6
tabulate==0.10.0
pyyaml==6.0.3
tqdm==4.67.3

# ============================================
# AI/ML - FILE INTELLIGENCE
# ============================================
scikit-learn==1.8.0

# ============================================
# DATA PROCESSING
# ============================================
pandas==3.0.1
openpyxl==3.1.5
pypdf==6.9.1
python-docx==1.2.0
chardet==7.2.0

# ============================================
# SYSTEM MONITORING
# ============================================
psutil==7.2.2
watchdog==6.0.0

# ============================================
# TIME SERIES & ANALYSIS
# ============================================
statsmodels==0.14.6
numpy==2.4.3
scipy==1.17.1

# ============================================
# UTILITIES
# ============================================
python-dateutil==2.9.0.post0
requests==2.32.5
